Added Like/Dislike/Flag with Posts (#4)

+++ nameless-webapp/persistence/src/main/java/ar/com/nameless/persistence/PostHibDao.java
@@ -1,23 +1,24 @@
 package ar.com.nameless.persistence;
 
 import ar.com.nameless.interfaces.dao.PostDao;
+import ar.com.nameless.model.FlaggedPost;
 import ar.com.nameless.model.FreshPost;
 import ar.com.nameless.model.HotPost;
 import ar.com.nameless.model.Post;
-import ar.com.nameless.model.Tag;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
-import javax.persistence.criteria.*;
+import javax.persistence.criteria.CriteriaBuilder;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Expression;
+import javax.persistence.criteria.Root;
 import java.util.Arrays;
 import java.util.List;
 
-/**
- * Created by root on 1/16/17.
- */
+
 @Repository
 public class PostHibDao implements PostDao{
 
@@ -38,6 +39,39 @@ public class PostHibDao implements PostDao{
         HotPost hotPost = new HotPost(post);
         entityManager.persist(hotPost);
         return hotPost;
+    }
+
+    @Transactional
+    public FlaggedPost newFlaggedPost(Post post) {
+        FlaggedPost flaggedPost = new FlaggedPost(post);
+        entityManager.persist(flaggedPost);
+        return flaggedPost;
+    }
+
+    @Transactional
+    public Post updateRating(long id, int delta){
+        Post post = findById(id);
+        if(post == null) return null;
+
+        post.setRating(post.getRating()+delta);
+        return post;
+    }
+
+    @Transactional
+    public Post flagPost(long id, int delta){
+        Post post = updateRating(id, delta);
+        if(post != null){
+            post.setFlags(post.getFlags()+1);
+        }
+        return post;
+    }
+
+    public boolean isFreshPost(Post post){
+        return findFreshPostByPost(post) != null;
+    }
+
+    public boolean isHotPost(Post post){
+        return findHotPostByPost(post) != null;
     }
 
     public Post findById(long id) {
@@ -85,8 +119,7 @@ public class PostHibDao implements PostDao{
     }
 
     @Transactional
-    public boolean deletePost(long id) {
-        //TODO: ojo que romperia al mover uno de fresh a hot
+    public boolean removeFromHot(long id){
         Post post = findById(id);
         if(post == null) return false;
 
@@ -96,6 +129,14 @@ public class PostHibDao implements PostDao{
             return true;
         }
 
+        return false;
+    }
+
+    @Transactional
+    public boolean removeFromFresh(long id){
+        Post post = findById(id);
+        if(post == null) return false;
+
         FreshPost freshPost = findFreshPostByPost(post);
         if(freshPost != null){
             entityManager.remove(freshPost);
@@ -104,6 +145,7 @@ public class PostHibDao implements PostDao{
 
         return false;
     }
+
 
     private HotPost findHotPostByPost(Post post) {
         final TypedQuery<HotPost> query = entityManager.createQuery("from HotPost as hp where hp.post= :post", HotPost.class);
